+++ packages/salesforce-adapter/src/types.ts
@@ -159,7 +159,7 @@ export type ChangeValidatorName =
   | 'deletedNonQueryableFields'
   | 'instanceWithUnknownType'
   | 'artificialTypes'
-  | 'deployNonDeployableTypes'
+  | 'metadataTypes'
 
 type ChangeValidatorConfig = Partial<Record<ChangeValidatorName, boolean>>
 
@@ -874,7 +874,7 @@ const changeValidatorConfigType =
       deletedNonQueryableFields: { refType: BuiltinTypes.BOOLEAN },
       instanceWithUnknownType: { refType: BuiltinTypes.BOOLEAN },
       artificialTypes: { refType: BuiltinTypes.BOOLEAN },
-      deployNonDeployableTypes: { refType: BuiltinTypes.BOOLEAN },
+      metadataTypes: { refType: BuiltinTypes.BOOLEAN },
     },
     annotations: {
       [CORE_ANNOTATIONS.ADDITIONAL_PROPERTIES]: false,
